Allow function calls in expressions
Allow function calls to be generated as part of expressions. For instance, if f: (x: Int, y: Int) -> Int, allows expressions like f(x: 5, y: 6) to be generated. The implementation of this involves significant changes to the way that expressions are generated.

Previously, the SwiftSmith grammar had a specific sub-grammar for constructing expressions of type `Int` and `Bool`. For instance, it included productions like
* EXPRESSION <Int> -> Variable<Int> " + " EXPRESSION<Int>
* EXPRESSION<Bool> -> EXPRESSION<Int> " > " EXPRESSION<Int>

This commit unifies this with other function call syntax (previously unsupported) by modeling + as an infix function of two arguments. This is similar to the way that the Swift compiler represents all operators. Now, an expression is produced with this algorithm:

1. decide whether to produce an atomic expression (variable, literal, or enum case), or a function call. If it is a variable or literal/enum case, then construct it and return.
1. construct a list of the types that are accessible in a scope
2. construct a list of the functions that are accessible in a scope. Currently, this includes only ordinary functions and static methods on types that are in scope.
3. generate a function call using the function name and atomic expressions to populate its arguments.

---

Also, fixed bug where functions were added to scope before they were defined. This allowed generation of infinitely recursive programs.