Fix broken tests, update docs
I broke a few tests when I changed the return type of `Scope.accessible_types` and `Scope.accessible_variables` from `set` to `list` as part of commit 88f96cf. This was done to ensure deterministic program generation, but the tests were not updated.

This also seemed like a good opportunity to update a few things in the README.