Eliminate deepcopy of tree
When applying an MR to a parse tree, the parse tree was previously copied. This is an expensive operation because the tree is deeply recursive. For particularly large trees, it could also cause a recursion depth error.

Instead, we modify the single tree after the first version of the program has been written.