Allow tests to be generated in isolated subdirectory
Previously, the run script generated, built, and ran tests in a single fixed directory called `generated`.  Now tests are generated, built, and run in a configurable directory. This is to set up for running multiple instances of SwiftSmith in parallel to accelerate testing.

This was also an opportunity to clean up an ill-thought-out compilation process formerly defined in `generate_module.sh`. By changing the working directory to the generation directory, cleaning up files in the project root directory was no longer necessary and the script could be eliminated.