Improvements to run on ARM
Previously, the test runner changed the path of shared libraries in the test executable using the MacOS-only install_name_tool. This is more complicated and less portable than simply setting the LD_LIBRARY_PATH environment variable to the working directory when running the tests. Changing this allows SwiftSmith to be run on my Raspberry Pi 3B+ which runs Linux and has an ARM processor.