Fix terrible bug in parallel test runner
Previously, the parallel test runner would "check" the result of the test execution. This raises an error if the program fails, which is not caught and terminates the thread. This has two terrible consequences:
1. The errors that we're looking for are not reported in the console.
2. The process stops running. Once an error happens on all processes, the runner stops doing any work.

This was also an opportunity to improve the run script: when compilation of one of the modules fails, that test run stops instead of inundating the console with lots of errors from the generated test script (because one of its imports doesn't exist).