Test runners log bug candidates to a file
Previously, any bugs encountered by the test runner would be printed out to the console. This makes it difficult to monitor a run over SSH. This commit solves this problem by writing any seeds which produced errors and the corresponding error messages to a log file called `bug_candidates.txt`.

Furthermore, the serial runner would stop running once a single bug candidate was detected, which is also resolved in the commit. This change was tested on both the serial and parallel runners.