Generate implicit member expressions for enum arguments to functions
In arguments to a function call, the types of the arguments can be inferred from the signature of the function (at least since protocols are not currently supported). In this situation, it is possible to use an implicit member expression, i.e. we can write `.a` instead of `A.a` where `A` is an enum with `case a`.

This will save some effort when calling functions in the generated test code, where there will be two versions of each generated type, one from `ModuleA` and one from `ModuleB`.